Implementation of DbSet.Local and data binding support
Issue #3915

Much of the implementation is the same as for EF6 and most of the tests have been ported directly. Key differences are:
- Calling .Local does not automatically call DetectChanges.
- .Local now returns a custom observable collection type rather than a type derived from ObservableCollection.

Both of these changes were driven by reported perf issues associated with the EF6 .Local. For the first point, DetectChanges can always be slow, and almost always .Local will return the same data regardless of whether DetectChanges is called.

For the second point, creating a new List containing a copy of all the tracked entities could be both slow and use a lot of memory. Also, the use of a list would mean inserts and removes from the list, which could be common, was slow. Therefore, the new type provides a view over the already tracked data without duplicating it.